+++ build_tools/get_tag.py
@@ -14,15 +14,16 @@ def get_version_from_tag(tag):
 
 
 # Circle is easy, since they give us the git tag
-if os.getenv('CIRCLECI', False) and os.getenv('GIT_TAG', False):
+if os.getenv('CIRCLECI', False) and os.getenv('CIRCLE_TAG', False):
     print('Tagged commit on Circle CI. Writing to {0}'.format(OUT_FILE))
     with open(OUT_FILE, 'w') as f:
-        tag = get_version_from_tag(os.getenv('GIT_TAG'))
+        tag = get_version_from_tag(os.getenv('CIRCLE_TAG'))
         f.write(tag)
 
 # on Azure Pipelines, we have to look at the build branch, and apply this logic:
 # refs/tags/vX.X.X -> ['refs', 'tags', 'vX.X.X'] -> 'vX.X.X'
-elif os.getenv('BUILD_SOURCEBRANCH', False) and os.getenv('BUILD_SOURCEBRANCH').startswith('refs/tags/'):
+elif os.getenv('BUILD_SOURCEBRANCH', False) and \
+        os.getenv('BUILD_SOURCEBRANCH').startswith('refs/tags/'):
     print('Tagged commit on Azure Pipelines. Writing to {0}'.format(OUT_FILE))
     with open(OUT_FILE, 'w') as f:
         tag = os.getenv('BUILD_SOURCEBRANCH').split('/')[-1]
@@ -30,4 +31,5 @@ elif os.getenv('BUILD_SOURCEBRANCH', False) and os.getenv('BUILD_SOURCEBRANCH').
 
 # Local or non-tagged commit, so we don't generate a VERSION file
 else:
-    print('Not a tagged commit, or not on a CI/CD platform. Not writing VERSION file')
+    print('Not a tagged commit, or not on a CI/CD platform. '
+          'Not writing VERSION file')