+++ examples/example_simple_fit.py
@@ -16,17 +16,18 @@ print(__doc__)
 # Author: Taylor Smith <taylor.smith@alkaline-ml.com>
 
 import pmdarima as pm
+from pmdarima import model_selection
 import numpy as np
 from matplotlib import pyplot as plt
 
 # #############################################################################
 # Load the data and split it into separate pieces
 data = pm.datasets.load_wineind()
-train, test = data[:150], data[150:]
+train, test = model_selection.train_test_split(data, train_size=150)
 
 # Fit a simple auto_arima model
-arima = pm.auto_arima(train, error_action='ignore', trace=1,
-                      suppress_warnings=True,
+arima = pm.auto_arima(train, error_action='ignore', trace=True,
+                      suppress_warnings=True, maxiter=10,
                       seasonal=True, m=12)
 
 # #############################################################################