+++ pmdarima/__init__.py
@@ -47,7 +47,9 @@ else:
 
     # Stuff we want at top-level
     from .arima import auto_arima, ARIMA, AutoARIMA, StepwiseContext, decompose
-    from .utils import acf, autocorr_plot, c, pacf, plot_acf, plot_pacf
+    from .utils import acf, autocorr_plot, c, pacf, plot_acf, plot_pacf, \
+        tsdisplay
+    from .utils._show_versions import show_versions
 
     # Need these namespaces at the top so they can be used like:
     # pm.datasets.load_wineind()
@@ -74,6 +76,7 @@ else:
         'pacf',
         'plot_acf',
         'plot_pacf',
+        'show_versions',
         'StepwiseContext',
     ]