+++ pmdarima/compat/__init__.py
@@ -1,14 +1,13 @@
 """
-The variables defined in compat are designed to provide
-compatibility between Python 2 & 3. Each sub-module is specifically
-designed not to make calls out to other portions of pmdarima and to
-remove circular dependencies.
+The variables defined in compat are designed to provide compatibility.
+Each sub-module is specifically designed not to make calls out
+to other portions of pmdarima and to remove circular dependencies.
 """
 
 from .matplotlib import *
 from .pandas import *
-from .python import *
 from .numpy import *
+from .sklearn import *
 from .statsmodels import *
 
 __all__ = [s for s in dir() if not s.startswith('_')]