+++ pmdarima/utils/array.py
@@ -4,10 +4,7 @@
 #
 # Array utilities
 
-from __future__ import absolute_import, division
-
 from sklearn.utils.validation import check_array, column_or_1d
-import six
 
 import numpy as np
 import pandas as pd
@@ -489,6 +486,6 @@ def is_iterable(x):
     isiter : bool
         True if iterable, else False.
     """
-    if isinstance(x, six.string_types):
+    if isinstance(x, str):
         return False
     return hasattr(x, '__iter__')