+++ pmdarima/utils/tests/test_meta.py
@@ -1,6 +1,4 @@
 
-from __future__ import absolute_import, print_function
-
 from pmdarima.utils.metaestimators import if_has_delegate